feat: reorganiza arquivos relacionados ao bg image e escreve alguns testes
