+++ .env.defaults
@@ -1,3 +1,4 @@
 NODE_ENV='development'
 PORT='3000'
 APP_URL=http://localhost:3000
+BING_API_URL=https://www.bing.com/HPImageArchive.aspx?format=js&idx=0&n=1&mkt=pt-BR
