+++ .gitignore
@@ -23,4 +23,4 @@ chrome-user-data
 .vscode
 *.swp
 *.swo
-.tmp
+.tmp/
