+++ .nvmrc
@@ -1,0 +1,1 @@
+v18.16.0
