+++ jest-setup.ts
@@ -1,0 +1,4 @@
+import "@testing-library/jest-dom";
+
+console.log = () => {}; //silence, i'll kill ya
+console.error = () => {};
