+++ package.json
@@ -28,6 +28,8 @@
         "@swc/jest": "^0.2.24",
         "@swc/plugin-styled-components": "^1.5.58",
         "@swc/register": "^0.1.10",
+        "@testing-library/jest-dom": "^5.16.5",
+        "@testing-library/react": "^14.0.0",
         "@types/express": "^4.17.13",
         "@types/jest": "^29.4.4",
         "@types/jsdom": "^21.1.0",
@@ -45,6 +47,7 @@
         "dotenv": "^16.0.1",
         "dotenv-defaults": "^5.0.0",
         "jest": "^29.4.3",
+        "jest-environment-jsdom": "^29.5.0",
         "jsdom": "^21.1.1",
         "nodemon": "^2.0.16",
         "prettier": "2.8.4",
