+++ react/hooks/use-imageurl.ts
@@ -1,0 +1,19 @@
+import { useQuery } from "@tanstack/react-query";
+import { BING_BASE_URL } from "../../shared/constants/index";
+
+interface ImgProps {
+    src: string;
+}
+
+export default function useImageUrl(): ImgProps {
+    const { data } = useQuery({
+        queryKey: ["background-image"],
+        queryFn: async () => {
+            return fetch(
+                `${window.isServer ? process.env.APP_URL : ""}/get-image`
+            ).then((res) => res.json());
+        },
+        suspense: true,
+    });
+    return { src: `${BING_BASE_URL}${data.url}` };
+}
