+++ shared/interfaces/FeaturedImageData.ts
@@ -1,0 +1,3 @@
+export interface FeaturedImageData {
+    url: string;
+}
